requirements.txt
pandas==2.2.0
numpy==1.26.3
pyyaml==6.0
torch==2.1.2
matplotlib==3.8.2
scipy==1.11.4
simpleitk==2.3.1
opencv-python==4.9.0.80
monai==1.4.0
json5==0.9.14
monai==1.4.0